+++ cstree/src/interning/lasso_compat/token_interner.rs
@@ -86,8 +86,8 @@ mod multi_threaded {
 
     /// A threadsafe [`Interner`] for deduplicating [`GreenToken`](crate::green::GreenToken) strings.
     ///
-    /// Note that [`Interner`] and [`Resolver`] are also implemented for  `&MultiThreadTokenInterner` so you can pass
-    /// `&mut &interner` in shared contexts.
+    /// Note that [`Interner`] and [`Resolver`] are also implemented for `Arc<MultiThreadTokenInterner>` so you can pass
+    /// `&mut Arc::clone(interner)` in shared contexts.
     #[cfg_attr(doc_cfg, doc(cfg(feature = "multi_threaded_interning")))]
     #[derive(Debug)]
     pub struct MultiThreadedTokenInterner {
@@ -106,7 +106,5 @@ mod multi_threaded {
     }
 
     impl_traits!(for MultiThreadedTokenInterner, if #[cfg(feature = "multi_threaded_interning")]);
-
-    impl_traits!(for &MultiThreadedTokenInterner, if #[cfg(feature = "multi_threaded_interning")]);
     impl_traits!(for StdArc<MultiThreadedTokenInterner>, if #[cfg(feature = "multi_threaded_interning")]);
 }